requirements.txt
pandas
yfinance
openai
requests
ta
python-dotenv
tqdm
numpy
